begin move of transport properties
